fix remote script path, fix mqtt stability, add focus script
- see docstring of new focus script. It uses the variance calculation
from Pillow, and probably doesn't work if the FOV is changing much. Be
careful it's easy to get caught in a local min as well.
- mqtt now takes into account if a 0 is put on the STAT channel. When
the 1 is received again, signifying it's back up, it will resend the
last packet and continue waiting. It used to wait forever for a DONE
that would never come.